Put together all line validations in PT SAFT
