+++ addons/pt/saft/line.go
@@ -7,6 +7,7 @@ import (
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/i18n"
+	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
 	"github.com/invopop/validation"
@@ -52,10 +53,29 @@ func validateLine(line *bill.Line) error {
 	}
 
 	return validation.ValidateStruct(line,
+		validation.Field(&line.Sum, num.ZeroOrPositive),
+		validation.Field(&line.Total, num.ZeroOrPositive),
+		validation.Field(&line.Discounts,
+			validation.Each(
+				validation.By(validateInvoiceLineDiscount),
+				validation.Skip,
+			),
+		),
 		validation.Field(&line.Notes,
 			validation.By(validateLineNotes(line)),
 			validation.Skip,
 		),
+	)
+}
+
+func validateInvoiceLineDiscount(val any) error {
+	disc, _ := val.(*bill.LineDiscount)
+	if disc == nil {
+		return nil
+	}
+
+	return validation.ValidateStruct(disc,
+		validation.Field(&disc.Amount, num.ZeroOrPositive),
 	)
 }
